fix .odbc.ini file handling (#55)

+++ docker/docker-entrypoint.sh
@@ -43,7 +43,7 @@ if [[ "$1" == "irods-start" ]]; then
             cp /etc/irods/irods_environment.json /var/lib/irods/.irods/irods_environment.json
         fi
 
-        if [ ! -f /var/lib/irods/.odbc.ini ]; then
+        if [[ -f /etc/irods/.odbc.ini ] && [ ! -f /var/lib/irods/.odbc.ini ]]; then
             cp /etc/irods/.odbc.ini /var/lib/irods/.odbc.ini
         fi
 
@@ -79,14 +79,17 @@ if [[ "$1" == "irods-start" ]]; then
         python3 /var/lib/irods/scripts/setup_irods.py --json_configuration_file=/unattended_config.json
 
         cp /var/lib/irods/.irods/irods_environment.json /etc/irods/irods_environment.json
-        cp /var/lib/irods/.odbc.ini /etc/irods/.odbc.ini
+
+        if [ -f /var/lib/irods/.odbc.ini ]; then
+            cp /var/lib/irods/.odbc.ini /etc/irods/.odbc.ini
+        fi
         cp -f /var/lib/irods/version.json /etc/irods/version.json
 
         touch /etc/irods/.provisioned
 
     fi
 
-    if [[ "$IRODS_ROLE" == "provider" ]]; then
+    if [ "$IRODS_ROLE" == "provider" ]; then
 
         echo "Set up custom PAM module"
         mkdir -p /usr/local/lib/pam-sodar
